requirements-dev.txt
django-extensions==3.2.3
django-cors-headers==4.2.0
flake8==6.1.0
isort==5.12.0
flake8-blind-except==0.2.1
flake8-broken-line==1.0.0
flake8-bugbear==23.7.10
flake8-builtins==2.1.0
flake8-class-attributes-order==0.1.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-eradicate==1.5.0
flake8-functions==0.0.8
flake8-isort==6.0.0
flake8-mutable==1.2.0
flake8-print==5.0.0
flake8-pytest==1.4
flake8-pytest-style==1.7.2
flake8-quotes==3.3.2
flake8-string-format==0.3.0
flake8-variables-names==0.0.6
pre-commit==3.3.3
pytest==7.4.0
wemake-python-styleguide==0.18.0